requirements.txt
Django>=5.2.7
django-simple-history>=3.5.0